+++ src/interfaces/CollectionInterface.ts
@@ -1,12 +1,10 @@
-import ResourceInterface from './ResourceInterface'
-import { StoreDataCollection } from './StoreData'
+import type ResourceInterface from './ResourceInterface'
 
-interface CollectionInterface extends ResourceInterface {
-    _storeData: StoreDataCollection
-
-    items: Array<ResourceInterface>
-    allItems: Array<ResourceInterface>
-    $loadItems: () => Promise<CollectionInterface>
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export interface CollectionInterface<ItemType extends ResourceInterface, ResourceType extends CollectionInterface<ItemType, ResourceType> = any> extends ResourceInterface<ResourceType> {
+    items: Array<ItemType>
+    allItems: Array<ItemType>
+    $loadItems: () => Promise<CollectionInterface<ItemType, ResourceType>>
 }
 
 export default CollectionInterface